soroban_ledger_snapshot
Struct LedgerSnapshot

pub struct LedgerSnapshot {
    pub protocol_version: u32,
    pub sequence_number: u32,
    pub timestamp: u64,
    pub network_id: [u8; 32],
    pub base_reserve: u32,
    pub min_persistent_entry_ttl: u32,
    pub min_temp_entry_ttl: u32,
    pub max_entry_ttl: u32,
    pub ledger_entries: Vec<(Box<LedgerKey>, (Box<LedgerEntry>, Option<u32>))>,
}

Ledger snapshot stores a snapshot of a ledger that can be restored for use in environments as a LedgerInfo and a SnapshotSource.

Fields
protocol_version: u32sequence_number: u32timestamp: u64network_id: [u8; 32]base_reserve: u32min_persistent_entry_ttl: u32min_temp_entry_ttl: u32max_entry_ttl: u32ledger_entries: Vec<(Box<LedgerKey>, (Box<LedgerEntry>, Option<u32>))>
Implementations

impl LedgerSnapshot

pub fn from<'a>( info: LedgerInfo, entries: impl IntoIterator<Item = (&'a Box<LedgerKey>, (&'a Box<LedgerEntry>, Option<u32>))>, ) -> Self

pub fn update(&mut self, host: &Host)

Update the snapshot with the state within the given Host.

The ledger info of the host will overwrite the ledger info in the snapshot. The entries in the host’s storage will overwrite entries in the snapshot. Existing entries in the snapshot that are untouched by the host will remain.

pub fn ledger_info(&self) -> LedgerInfo

pub fn set_ledger_info(&mut self, info: LedgerInfo)

Set the ledger info in the snapshot.

pub fn entries( &self, ) -> impl IntoIterator<Item = (&Box<LedgerKey>, &(Box<LedgerEntry>, Option<u32>))>

Get the entries in the snapshot.

pub fn set_entries<'a>( &mut self, entries: impl IntoIterator<Item = (&'a Box<LedgerKey>, (&'a Box<LedgerEntry>, Option<u32>))>, )

Replace the entries in the snapshot with the entries in the iterator.

pub fn update_entries<'a>( &mut self, entries: impl IntoIterator<Item = &'a (Rc<LedgerKey>, Option<(Rc<LedgerEntry>, Option<u32>)>)>, )

Update entries in the snapshot by adding or replacing any entries that have entry in the input iterator, or removing any that does not have an entry.

impl LedgerSnapshot

pub fn read(r: impl Read) -> Result<LedgerSnapshot, Error>

pub fn read_file(p: impl AsRef<Path>) -> Result<LedgerSnapshot, Error>

pub fn write(&self, w: impl Write) -> Result<(), Error>

pub fn write_file(&self, p: impl AsRef<Path>) -> Result<(), Error>